+++ Project3/WickedDoge.h
@@ -9,9 +9,9 @@ private:
 
 	// status
 	//time since the last time the entity took damage
-	double timeSinceHurt = 0.f;
+	float timeSinceHurt = 0.f;
 	//minimum time between taking damage
-	double minHurtTimeGap = 2.f;
+	float minHurtTimeGap = 2.f;
 
 
 	//initializer functions
@@ -19,8 +19,8 @@ private:
 	void initComponents();
 
 	// implement flicker
-	const double flickerTimeGap = 0.1f;
-	double currentFlickerTime = 0.f;
+	const float flickerTimeGap = 0.1f;
+	float currentFlickerTime = 0.f;
 	bool flickerNoRender;
 
 	float currenDt = 0.f;
